change file index.js

+++ src/index.js
@@ -40,10 +40,12 @@ cron.schedule('0 * */3 * * *', async () => {
     console.log('Memulai Sinkronisasi data', `${time.getHours()} : ${time.getMinutes()} : ${time.getSeconds()}`)
    
     const waterUsage = await api.getWaterUsages();
-    await database('waterusage').truncate()
-    await database('waterusage').insert(waterUsage);
-
-    console.log('Sinkronisasi Data berhasil', `${time.getHours()} : ${time.getMinutes()} : ${time.getSeconds()}`)
+    
+    if (waterUsage.length > 0) {
+      await database('waterusage').truncate()
+      await database('waterusage').insert(waterUsage);
+      console.log('Sinkronisasi Data berhasil', `${time.getHours()} : ${time.getMinutes()} : ${time.getSeconds()}`)
+    }
   } catch (error) {
     console.log(error)
   }
